Remove `ZoomHub.Types.Internal` namespace
