+++ src/ZoomHub/Types/EmbedId.hs
@@ -6,14 +6,14 @@ module ZoomHub.Types.EmbedId
   , unEmbedId
   ) where
 
-import qualified Data.ByteString.Char8            as BC
-import qualified Data.Text                        as T
-import           Network.HTTP.Types               (decodePath)
-import           Servant                          (FromText, fromText)
-import           System.FilePath.Posix            (splitExtension)
+import qualified Data.ByteString.Char8   as BC
+import qualified Data.Text               as T
+import           Network.HTTP.Types      (decodePath)
+import           Servant                 (FromText, fromText)
+import           System.FilePath.Posix   (splitExtension)
 
-import           ZoomHub.Types.Internal.ContentId (ContentId)
-import qualified ZoomHub.Types.Internal.ContentId as ContentId
+import           ZoomHub.Types.ContentId (ContentId)
+import qualified ZoomHub.Types.ContentId as ContentId
 
 newtype EmbedId = EmbedId { unEmbedId :: ContentId } deriving (Eq, Show)
