+++ src/ZoomHub/Types/ViewContent.hs
@@ -7,21 +7,18 @@ module ZoomHub.Types.ViewContent
   )
   where
 
-import           Data.Maybe                       (fromJust)
-import           Data.Monoid                      ((<>))
-import qualified Data.Text                        as T
-import           Lucid                            (ToHtml, body_, charset_,
-                                                   content_, doctypehtml_,
-                                                   head_, href_, link_, meta_,
-                                                   name_, rel_, script_, src_,
-                                                   style_, title_, toHtml,
-                                                   toHtmlRaw)
-import           Network.URI                      (parseRelativeReference,
-                                                   relativeTo)
+import           Data.Maybe                (fromJust)
+import           Data.Monoid               ((<>))
+import qualified Data.Text                 as T
+import           Lucid                     (ToHtml, body_, charset_, content_,
+                                            doctypehtml_, head_, href_, link_,
+                                            meta_, name_, rel_, script_, src_,
+                                            style_, title_, toHtml, toHtmlRaw)
+import           Network.URI               (parseRelativeReference, relativeTo)
 
-import           ZoomHub.API.Types.Content        (Content, contentId)
-import           ZoomHub.Types.BaseURI            (BaseURI, unBaseURI)
-import           ZoomHub.Types.Internal.ContentId (unId)
+import           ZoomHub.API.Types.Content (Content, contentId)
+import           ZoomHub.Types.BaseURI     (BaseURI, unBaseURI)
+import           ZoomHub.Types.ContentId   (unId)
 
 data ViewContent = ViewContent
   { vcContent :: Content
